requirements.txt
requests
tldextract
python-whois
beautifulsoup4
dnspython
Flask
jsonify